requirements.txt
faiss-cpu==1.7.4
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.1
langchain==0.2.17
numpy==1.24.3
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
sentence-transformers
transformers
huggingface_hub
python-dotenv